Break up graphFrame
