+++ gui/shipBrowser.py
@@ -1958,3 +1958,21 @@ class FitItem(SFItem.SFBrowserItem):
         self.bkBitmap.sFactor = sFactor
         self.bkBitmap.eFactor = eFactor
         self.bkBitmap.mFactor = mFactor
+
+
+class FitList(wx.Panel):
+    def __init__(self, parent):
+        wx.Panel.__init__(self, parent)
+        self.mainSizer = wx.BoxSizer(wx.VERTICAL)
+        self.SetSizer(self.mainSizer)
+
+        self.fitList = FitDisplay(self)
+        self.mainSizer.Add(self.fitList, 1, wx.EXPAND)
+        fitToolTip = wx.ToolTip("Drag a fit into this list to graph it")
+        self.fitList.SetToolTip(fitToolTip)
+
+
+    DEFAULT_COLS = ["Base Icon",
+                    "Base Name"]
+
+    def __init__(self, parent):
